devided keys class  component from calculator component
